api/requirements.txt
fastapi==0.93.0
uvicorn==0.20.0
python-dotenv==1.0.0
